pyproject.toml
[tool.poetry]
name = "fastapi-recipes-project"
version = "0.1.0"
description = "Backend API for managing recipes using FastAPI"
authors = ["Juan Facundo Quallio <facuq93@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
fastapi = "^0.110"
uvicorn = "^0.29"
sqlalchemy = "^2.0"
psycopg2-binary = "^2.9"
python-dotenv = "^1.0"
pydantic = {extras = ["email"], version = "^2.11.7"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
